requirements.txt
fastapi>=0.100
uvicorn[standard]>=0.20
psycopg2-binary>=2.9
redis>=4.0
bcrypt>=4.0
PyJWT>=2.8
pydantic[email]>=2.0
